prov/rxm: Move msg ops into its own file
Move the message ops and related static functions into rxm_msg.c
out of rxm_ep.c.  This makes the functionality easier to locate.

Minor renames and white space adjustments.  No functional changes
are made to the code.
