+++ libfabric.vcxproj
@@ -372,6 +372,7 @@
     </ClCompile>
     <ClCompile Include="prov\rxm\src\rxm_attr.c" />
     <ClCompile Include="prov\rxm\src\rxm_conn.c" />
+    <ClCompile Include="prov\rxm\src\rxm_msg.c" />
     <ClCompile Include="prov\rxm\src\rxm_rma.c" />
     <ClCompile Include="prov\rxm\src\rxm_cq.c" />
     <ClCompile Include="prov\rxm\src\rxm_av.c" />
