+++ libfabric.vcxproj.filters
@@ -360,6 +360,9 @@
     <ClCompile Include="prov\rxm\src\rxm_rma.c">
       <Filter>Source Files\prov\rxm\src</Filter>
     </ClCompile>
+    <ClCompile Include="prov\rxm\src\rxm_msg.c">
+      <Filter>Source Files\prov\rxm\src</Filter>
+    </ClCompile>
     <ClCompile Include="src\iov.c">
       <Filter>Source Files\src</Filter>
     </ClCompile>
